+++ packages/react/src/experimental/Information/Headers/BaseHeader/index.tsx
@@ -57,6 +57,7 @@ interface BaseHeaderProps {
     actions?: MetadataAction[]
   }
   metadata?: MetadataProps["items"]
+  metadataRowGap?: MetadataProps["rowGap"]
   /** Renders a 1px bottom border at the very bottom of the header. */
   showBottomBorder?: boolean
 }
@@ -74,6 +75,7 @@ export function BaseHeader({
   otherActions = [],
   status,
   metadata = [],
+  metadataRowGap = "none",
   showBottomBorder = false,
 }: BaseHeaderProps) {
   const allMetadata: BaseHeaderProps["metadata"] = [
@@ -168,7 +170,7 @@ export function BaseHeader({
 
         {allMetadata.length > 0 && (
           <div className="flex flex-wrap items-center gap-x-3 gap-y-1 md:hidden">
-            <Metadata items={allMetadata} />
+            <Metadata items={allMetadata} rowGap={metadataRowGap} />
           </div>
         )}
 
@@ -310,7 +312,7 @@ export function BaseHeader({
       </div>
       {allMetadata.length > 0 && (
         <div className="hidden flex-wrap items-center gap-x-3 gap-y-1 md:block">
-          <Metadata items={allMetadata} />
+          <Metadata items={allMetadata} rowGap={metadataRowGap} />
         </div>
       )}
     </div>