+++ packages/react/src/lib/exports.ts
@@ -11,6 +11,18 @@ export { experimentalComponent as experimental } from "./experimental"
 export { OneEllipsis } from "./OneEllipsis"
 export { PrivacyModeProvider, usePrivacyMode } from "./privacyMode"
 export * from "./providers/datacollection/exports"
+export { dialog, drawer } from "./providers/dialogs-alike"
+export type {
+  DialogDefinition,
+  DialogActions,
+  DialogAction,
+  DialogActionValue,
+  DialogId,
+  DrawerDefinition,
+  NotificationDialogOptions,
+  AlertDialogOptions,
+  ConfirmDialogOptions,
+} from "./providers/dialogs-alike"
 export { F0EventCatcherProvider } from "./providers/events"
 export { buildTranslations } from "./providers/i18n"
 export type { TranslationsType } from "./providers/i18n"